kernel: Linux IP Virtual Server kernel modules (netfilter.mk patch)
IPVS (IP Virtual Server) implements transport-layer load balancing inside the Linux kernel
so called Layer-4 switching.
IPVS running on a host acts as a load balancer at the front of a cluster of real servers, it
can direct requests for TCP/UDP based services to the real servers, and makes services of
the real servers to appear as a virtual service on a single IP address.

Signed-off-by: Mauro Mozzarelli <mauro@ezplanet.net>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@48982 3c298f89-4303-0410-b956-a3cf2f4a3e73